dating_app/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pillow==9.5.0
celery==5.3.0
django-filter==23.2
django-cors-headers==4.0.0
PyYAML==6.0
uritemplate==4.1.1
gunicorn==20.1.0